requirements.txt
```txt
python-telegram-bot==20.7
requests==2.31.0
python-decouple==3.8
sqlite3